requirements.txt
discord.py
requests
gitpython
